combine ignore schema traversal with rule traversal
prior to this commit the schema was first traversed
to find keywords to blend then traversed again to find
keywords to ignore (which were only array/ListNode)
items. With this commit the 2 traversals are combined
where the array/ListNode items are skipped during
the finding traversal instead of found then ignored.

The unit tests are also updated to test that this
combination works (and to test blendmodels, the
only "public" API for this submodule).